LG-2556: adds helper methods for signing in from an oidc sp
