+++ spec/features/openid_connect/authorization_confirmation_spec.rb
@@ -1,6 +1,8 @@
 require 'rails_helper'
 
 feature 'OIDC Authorization Confirmation' do
+  include OidcAuthHelper
+
   before do
     allow(Figaro.env).to receive(:otp_delivery_blocklist_maxretry).and_return('9999')
   end
@@ -31,7 +33,7 @@ feature 'OIDC Authorization Confirmation' do
 
     it 'it confirms the user wants to continue to the SP after signing in again' do
       sign_in_user(user1)
-      visit_idp_from_oidc_sp
+      visit_idp_from_ial1_oidc_sp
       expect(current_url).to match(user_authorization_confirmation_path)
 
       continue_as(user1.email)
@@ -40,7 +42,7 @@ feature 'OIDC Authorization Confirmation' do
 
     it 'it allows the user to switch accounts prior to continuing to the SP' do
       sign_in_user(user1)
-      visit_idp_from_oidc_sp
+      visit_idp_from_ial1_oidc_sp
       expect(current_url).to match(user_authorization_confirmation_path)
 
       continue_as(user2.email, user2.password)
@@ -54,7 +56,7 @@ feature 'OIDC Authorization Confirmation' do
 
     it 'does not render an error if a user goes back after opting to switch accounts' do
       sign_in_user(user1)
-      visit_idp_from_oidc_sp
+      visit_idp_from_ial1_oidc_sp
 
       expect(current_path).to eq(user_authorization_confirmation_path)
 
@@ -64,28 +66,5 @@ feature 'OIDC Authorization Confirmation' do
 
       expect(current_path).to eq(new_user_session_path)
     end
-  end
-
-  def sign_in_oidc_user(user)
-    visit_idp_from_oidc_sp
-    fill_in_credentials_and_submit(user.email, user.password)
-    click_continue
-  end
-
-  def visit_idp_from_oidc_sp
-    client_id = 'urn:gov:gsa:openidconnect:sp:server'
-    state = SecureRandom.hex
-    nonce = SecureRandom.hex
-
-    params = {
-      client_id: client_id,
-      response_type: 'code',
-      acr_values: Saml::Idp::Constants::IAL1_AUTHN_CONTEXT_CLASSREF,
-      scope: 'openid email',
-      redirect_uri: 'http://localhost:7654/auth/result',
-      state: state,
-      nonce: nonce,
-    }
-    visit openid_connect_authorize_path(params)
   end
 end
